Refine the logic of open and create page store
